Create All Unique Permutations of an array.java
Boost your resume to top companies via this Free Hiring Contest! 

banner
Given an array arr[] of length n. Find all possible unique permutations of the array in sorted order. A sequence A is greater than sequence B if there is an index i for which Aj = Bj for all j<i and Ai > Bi.